1789: refactor the category header row into the GbHeaderToolbar component so it uses <th>s and is rendered in the correct position (doesn't require any javascript tomfoolery)
