+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/component/GbDataTable.java
@@ -1,7 +1,5 @@
 package org.sakaiproject.gradebookng.tool.component;
 
-import org.apache.wicket.AttributeModifier;
-import org.apache.wicket.extensions.markup.html.repeater.data.sort.ISortStateLocator;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.AbstractColumn;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.DataTable;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.IColumn;
@@ -11,16 +9,21 @@ import org.apache.wicket.markup.repeater.Item;
 import org.apache.wicket.model.IModel;
 
 import java.util.List;
+import java.util.Map;
 
 public class GbDataTable<T, S> extends DataTable<T, S> {
+
+	IModel<Map<String, Object>> gradebookData;
+
 	public GbDataTable(final String id, final List<? extends IColumn<T, S>> columns,
-										 final IDataProvider<T> dataProvider, final long rowsPerPage) {
+			final IDataProvider<T> dataProvider, final long rowsPerPage, final IModel<Map<String, Object>> model) {
 		super(id, columns, dataProvider, rowsPerPage);
+		this.gradebookData = model;
 	}
 
+
 	@Override
-	protected Item<IColumn<T, S>> newCellItem(final String id, final int index,
-														 final IModel<IColumn<T, S>> model) {
+	protected Item<IColumn<T, S>> newCellItem(final String id, final int index, final IModel<IColumn<T, S>> model) {
 		return new Item(id, index, model) {
 			@Override
 			protected void onComponentTag(ComponentTag tag) {
@@ -38,5 +41,9 @@ public class GbDataTable<T, S> extends DataTable<T, S> {
 				}
 			}
 		};
+	}
+
+	public Map<String,Object> getGradebookData() {
+		return GbDataTable.this.gradebookData.getObject();
 	}
 }
