+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/pages/GradebookPage.html
@@ -41,7 +41,6 @@
       </ul>
     </div>
     <div id="gradebookGrades">
-      <div wicket:id="fixedHeader" id="gradebookFixedThings"></div>
       <div id="gradebookHorizontalOverflowWrapper">
           <div id="gradebookVerticalOverflowWrapper">
               <table wicket:id="table" id="gradebookGradesTable" class="table-striped table-bordered table-hover" role="grid"></table>
