--- gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/panels/GradebookSpreadsheetFixedTables.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns:wicket="http://wicket.apache.org/dtds.data/wicket-xhtml1.4-strict.dtd" >
-
-<body>
-<wicket:panel>
-
-	<table style="display: none;">
-		<thead>
-			<tr class="gb-categories-row">
-				<th colspan="3"></th>
-				<th wicket:id="categories">
-					<span class="gb-category-label">
-						<span wicket:id="extraCreditCategoryFlag" class="gb-category-extra-credit"></span>
-						<span wicket:id="name" class="gb-category-name"></span>
-						<span wicket:id="weight" class="gb-category-weight"></span>
-					</span>
-				</th>
-				<th class="gb-uncategorized" style="display: none;">
-					<span class="gb-category-label">
-						<span class="gb-category-name"><wicket:message key="gradebookpage.uncategorised"></wicket:message></span>
-					</span>
-				</th>
-			</tr>
-		</thead>
-	</table>
-</wicket:panel>
-</body>
-</html>
